requirements.txt
django-registration=1.0
django_facebook